+++ pat/discoproxyclient.py
@@ -65,11 +65,17 @@ class DiscoProxyClient:
     def __init__(
         self,
         disco_proxy_address: str,
-        my_address: str,
+        host_address: str,
         ssl_context_container: SSLContextContainer | None = None,
     ):
+        """Disco proxy client constructor.
+
+        :param disco_proxy_address: Address of the disco proxy, e.g. 127.0.0.1:9999
+        :param host_address: The address of this host. Used as wse:NotifyTo/@Address to subscribe to the disco proxy.
+        :param ssl_context_container: ssl configuration container. If None, no encryption is used.
+        """
         self._proxy_address = disco_proxy_address
-        self._my_address = my_address
+        self._host_address = host_address
         self._ssl_context_container = ssl_context_container
         self._logger = get_logger_adapter('sdc.disco')
         self._local_services: dict[str, Service] = {}
@@ -84,14 +90,13 @@ class DiscoProxyClient:
             msg_reader=message_reader,
         )
         self._http_server = HttpServerThreadBase(
-            my_address,
+            host_address,
             ssl_context_container.server_context if ssl_context_container else None,
             logger=get_logger_adapter('sdc.disco.httpsrv'),
             supported_encodings=['gzip'],
         )
 
         self._msg_converter = MessageConverterMiddleware(message_reader, message_factory, self._logger, self)
-        self._my_server_port = None
         self.subscribe_response = None
 
     def start(self, subscribe: bool = True, http_server_start_timeout: float = 60.0):
@@ -104,7 +109,6 @@ class DiscoProxyClient:
             msg = f'Http server could not be started within {http_server_start_timeout} seconds.'
             raise RuntimeError(msg)
         self._logger.info('Http server started. Serving EventSink on %s', self._http_server.base_url)
-        self._my_server_port = self._http_server.my_port
         self._http_server.dispatcher.register_instance('', self._msg_converter)
 
         if subscribe:
@@ -117,10 +121,10 @@ class DiscoProxyClient:
             self.send_unsubscribe()
         self._http_server.stop()
 
-    def get_active_addresses(self) -> list[str]:
+    @property
+    def active_address(self) -> str:
         """Get active addresses."""
-        # TODO: do not return list  # noqa: FIX002, TD002, TD003
-        return [self._my_address]
+        return self._host_address
 
     def search_services(
         self,
@@ -198,7 +202,8 @@ class DiscoProxyClient:
     def send_subscribe(self) -> ReceivedMessage:
         """Send subscribe message."""
         subscribe_request = eventing_types.Subscribe()
-        subscribe_request.Delivery.NotifyTo.Address = f'https://{self._my_address}:{self._my_server_port}'
+        assert self._http_server.server_port is not None, 'http server port is not set'
+        subscribe_request.Delivery.NotifyTo.Address = f'https://{self._host_address}:{self._http_server.server_port}'
         subscribe_request.Expires = 3600
         subscribe_request.set_filter('', dialect='http://discoproxy')
         inf = HeaderInformationBlock(action=subscribe_request.action, addr_to=ADDRESS_ALL)